text_2_sql/requirements.txt
numpy==1.24.4
pandas==2.0.3
pip==23.1.2
scikit-learn==1.3.0
SQLAlchemy==2.0.19
torch==2.0.1+cu118
transformers==4.32.0.dev0
sentencepiece=0.1.99
sentence-transformers==2.2.2
fire==0.5l0
Django==4.2.3
django-cors-headers=4.2.0
djangorestframework==3.14.0
bertopic==0.15.0
bitsandbytes==0.39.1
peft==0.5.0.dev0
protobuf==3.20.0
